pi-client/requirements.txt
websockets==12.*
sounddevice==0.4.*
numpy==1.*
python-dotenv==1.*
